==> 7-Deduplication/ODHF_copy/0_split_pharmacies.ipynb <==
import pandas as pd

# Make separate dataset of only pharmacies and covid-19 facilities

df = pd.read_csv('../7-Deduplication/inputs/Assigned_CSDs.csv')


df.loc[df['facility_name'].str.contains('hebergement', case=False, na=False), 'facility_type'] = 'Nursing and residential care facilities'
df.loc[df['facility_name'].str.contains('nursing station|centre de soins infirmier', case=False, na=False), 'facility_type'] = 'Ambulatory health care services'



df_phar = df.loc[df['facility_type'].isin(['Pharmacy','Covid-19 related facilities', 'Laboratory'])]
df_rest = df.loc[~df['facility_type'].isin(['Pharmacy','Covid-19 related facilities', 'Laboratory'])]

df_phar.reset_index(drop=True, inplace=True)

#df_rest = df_rest.dropna(subset=['longitude', 'latitude'], how='all')

df_phar.to_csv('../7-Deduplication/pharmacies_and_covid/inputs/pharmacies_only.csv', index=False)
df_rest.to_csv('../7-Deduplication/inputs/assigned_CSDs.csv', index=False)
